svn path=/tags/FastInfoset_1_1_6/; revision=693


Former-commit-id: 47fbfa0c7e3808df3961edcd6f98be6ef482ead4

+++ FastInfoset/src/com/sun/xml/fastinfoset/algorithm/DoubleEncodingAlgorithm.java
@@ -97,7 +97,7 @@ public class DoubleEncodingAlgorithm extends IEEE754FloatingPointEncodingAlgorit
                 new WordListener() {
             public void word(int start, int end) {
                 String fStringValue = cb.subSequence(start, end).toString();
-                doubleList.add(Float.valueOf(fStringValue));
+                doubleList.add(Double.valueOf(fStringValue));
             }
         }
         );
